+++ src/libs/mysql/traps.c
@@ -400,7 +400,7 @@ BARTLBY_SQL_PROTECTION_INIT;
     
 }
 
-int GetTrapMap(struct trap * svcs, char * config, int orch_id) {
+long GetTrapMap(struct trap * svcs, char * config, int orch_id) {
   
   MYSQL *mysql;
   MYSQL_ROW  row;
@@ -411,7 +411,7 @@ int GetTrapMap(struct trap * svcs, char * config, int orch_id) {
   char * mysql_user = getConfigValue("mysql_user", config);
   char * mysql_pw = getConfigValue("mysql_pw", config);
   char * mysql_db = getConfigValue("mysql_db", config);
-  int i=0;
+  long i=0;
   
   char * sql, *where;
